API NOVA

+++ app/src/main/java/br/com/analytics/educa/data/retrofit/ApiService.kt
@@ -54,6 +54,6 @@ data class ApiResponse(
 
 data class SchoolPerformance(
     val materia: String,
-    val media_nota: Double,
-    val media_presenca: Double
+    val media_nota: Float,
+    val media_presenca: Float
 )
